Fix websocket errors after reverting the commit that broke error parsing for the api.
The following commit fixed error parsing for websocket errors. But it
broke error parsing for the regular api errors. So I reverted the commit
and added the commit you are currently looking at to fix the websocket
errors.

Hopefully slack will unify the errors at some point and we can get rid
of one of them.

commit 2cfb1c0d1986ae8a8c4cfa37f316d29102e62de1
Author: Tarrant <tarrant@keyneston.com>
Date:   Fri Jan 30 00:00:43 2015 -0800

    Unmarshal SlackError as an error

    It looks like Slack changed how they return an error from a string
    to a
        hash.

    This meant that we could just unmarshal it straight into a
    struct. With
        a little helper function we automatically get go errors.