add support for setting and unsetting Custom Status/Emoji
This change introduces two new methods on *Client for setting and unsetting the
Custom Status and Status Emoji for the currently authenticated User. This also
adds `StatusText` and `StatusEmoji` fields to the `UserProfile` struct. The
Custom Status and Status Emoji functionality was recently added as part of the
User's profile[1].

The first method added is `SetUserCustomStatus()` which takes two string
arguments. This method uses an anonymous struct to generate the JSON required to
update a User's Custom Status and Status Emoji. If both arguments are set to
empty-string (`""`) the Slack API will unset the Custom Status/Emoji.

The second method is `UnsetUserCustomStatus()`, which is a convenience method
that takes zero arguments and wraps `SetUserCustomStatus()`. It unsets the
custom status.

Integration tests were added to ensure that that setting and unsetting the
Custom Status should work as-expected.

Closes #153.

[1] https://slackhq.com/set-your-status-in-slack-28a793914b98

Signed-off-by: Tim Heckman <t@heckman.io>