use pagination for users.list
Issue #196

Use pagination to try to get all users.
The API docs recommend to request 200 users each time, but it
there is also rate-limiting in effect, so if we loop too often
too quickly, we'll never get all users.  We try to work around
that by sleeping for 2 seconds in between each iteration (API
docs say "send no more than one message per second", so we
want to play it safe), but even then we get rate-limited when
trying to fetch several K users, so we try to get as many as
we can on each call (i.e. 999).

This is all very suboptimal, but getting, say, 8K users out of
20K is better than getting 0, so I suppose it's a start.