Improve usage of interface{} in the HTTP helper functions
The HTTP helper functions (post and postForm) take an `intf interface{}`
parameter, with the goal of using it to unmarshal the JSON response. This
requires that those calling the helper functions pass in a pointer to the type
they want to unmarshal the JSON in to.

The `parseResponseBody` function was a bit different in that it took a
`*interface{}` instead. Looking at its behaviors, this looks to have been
accidental when the unmarshaling behavior was implemented early on (1b96f659ad).
`json.Unmarshal` needs the `interface{}` value itself to contain a pointer, it
doesn't need to (and probably shouldn't) be a pointer.

To go in to deeper detail, the `interface{}` itself encapsulates the original
value and its type. When the JSON Unmarshaler tries to use the value within the
interface, it needs it to be a pointer so that it can manipulate the data and
set the values. The interface being a pointer only allows you to change what
data it encapsulates, and that's not generally how you'd use `interface{}` with
JSON.

This change updates the `*interface{}` typed argument of `parseResponseBody` to
be `interface{}` instead, and updates all of the call sites to use the new API.

Signed-off-by: Tim Heckman <t@heckman.io>