Remove omitempty on SelectBlockElement.MinQueryLength
In new Block API, default MinQueryLength is 3. It seems not to be documented;
however, it was confirmed to me by Slack support. It means that by default,
for an external_select, the query to populate the select menu will be sent
by Slack only if the user has typed 3 characters.

To change this value to 0 (in order to populate the select menu even if no
character is typed), setting MinQueryLength = 0 does not work because 0 is
Go default value for int and is omitted from JSON.

This change removes "omitempty" which means that setting MinQueryLength = 0
adds "min_query_length": 0 in JSON serialization. It also adds a default
value of 3.