Allow Fields and Accessory to be optional
According to the official Slack API documentation, Section blocks do not require fields and accessory argument to be required. 

The current implementation requires the user to specify both these fields if he/she wants to use the NewSectionBlock method.

In keeping with the spirit of the official implementation, I have made it such that these 2 arguments are now optional.