Issue 844: Redact api token in chat message
Debug logs are printing api tokens as part of the request message.
Instead of printing the token, the token is tarnsformed into a redacted
string if it is not empty. If the token is empty, which is a useful
piece of infomation that should be surfaced in the request, the empty
string is transformed into a nil string to make it more obvious.