Configure Dependabot to maintain GitHub Actions (#1374)
💁 When browsing this repository earlier, I saw that external actions are
referenced using commit SHAs. This is useful as a way of ensuring that
the version of a specific workflow doesn't drift unexpectedly but it
also means that they need to be maintained (whether actively or not).

This change configures @dependabot to do the hard work of looking after
those external action references and updating them as needed.