fix: Allow `emoji` value in plain text blocks to be null (#1354)
Fixes #1332.

This PR allows setting null values for the `emoji` field of plain text
blocks, which was previously incorrectly marshalling to JSON due to a
"false" value being omitted. I opted not to change the signature of
`NewTextBlockObject`, since the null value is only expected when
unmarshalling, and the user should know which value they need to use
when creating the block.