Fix message event struct
This has been the source of many bugs in our code.

Slack's docs say that a 'message event' contains a 'message' which is the same type as other representations of a message in slack.
However, the library has previously represented this as a nested MessageEvent which is recursively nested.
That creates lots of bugs, as you can never know if you're looking at the outer MessageEvent (which contains the event information) or the inner MessageEvent (which contains the message itself).

Instead, we can use our canonical Message instead.

Note that we also have to add a `root` onto Message, as this is populated for thread_broadcast events (for legacy reasons).