feat: add `ForceInviteUsersToConversation`
This is pretty much the same as `InviteUsersToConversation` but setting the param `force`
to true when calling the API.

I've done this to not break *even more* the API but this should have been a breaking
change to `InviteUsersToConversation` to use a params struct. I'm leaving that for a
future one.