Unmarshal messages correctly for all subtypes
When we get a 'message' event, the message details (blocks, attachments, etc.) are all included on the top level struct.
When it's a 'message_changed' event, these details are instead nested under a `message` key.
This is really really confusing!

We can't just nest a `Msg` struct inside our MessageEvent, because the keys of other message events
conflict with the `Msg` keys (e.g. both a `message` and a `message_changed` event will have a `user`
key).

To avoid this, we implement a custom unmarshaller which will:
1. If there's a `message` key in the payload, unmarshal it into the Message field
2. If there's no `message` key, populate the Message field with the event itself, assuming
   that it's an event with no subtype and so the message info is at the top level.

This means some information is represented twice in the same struct, but that is less bad
than it being unclear to a caller where they should go to access e.g. the message text.

Also reworks the tests so they use examples pulled directly from Slack's documentation.