feat: Align GetReactions with Slack API
!BREAKING CHANGE

Previously this method returned only reaction details, but with this change it started to also return the item itself (e.g. message, file, file_comment). This makes this closer to the actual response of Slack API.

To migrate an existing code, you can just use resp.Reactions to access
the slice of []ItemReaction.